refactor(layout): add `data-component` attributes and adjust container styles for improved structure and debugging in components
