+++ src/app/_components/inner-container.tsx
@@ -9,11 +9,11 @@ export function InnerContainer(props: InnerContainerProps) {
   const innerContainerStyles = container.raw({
     w: "full",
     px: { base: 5, md: 6 },
-    maxW: { base: "full", lg: "4xl" },
+    maxWidth: { base: "xl", sm: "2xl", lg: "4xl" },
   });
 
   return (
-    <div className={css(innerContainerStyles, props.styles)}>
+    <div className={css(innerContainerStyles, props.styles)} data-component="inner-container">
       {props.children}
     </div>
   );
